abstracted assert and CustomError into error.js
