+++ helpers/error.js
@@ -1,12 +1,3 @@
-function makeHttpError({ status = 500, message }) {
-
-	return {
-		success: false,
-		status,
-		errors: message
-	};
-}
-
 /**
  * Allows for an object to be passed as a message
  */
@@ -18,4 +9,32 @@ class MultipleError extends Error {
 	}
 }
 
-module.exports = { makeHttpError, MultipleError };
+//used to put errorcode in a context instead of just the message
+class CustomError extends Error {
+	constructor(errorCode) {
+		super(errorCode.message);
+		this.name = this.constructor.name;
+		this.code = errorCode.code;
+	}
+}
+
+
+function makeHttpError({ status = 500, message }) {
+	
+	return {
+		success: false,
+		status,
+		errors: message
+	};
+}
+
+//Asserts used for errorhandling
+function assert(condition, errorcode) {
+	if (!condition) {
+		throw new CustomError(errorcode);
+	}
+}
+
+
+
+module.exports = { assert, makeHttpError, MultipleError, CustomError };
